Add dynamic render to coa table
