--- src/components/ChartOfAccountsRow/index.tsx
@@ -1,1 +1,0 @@
-export { ChartOfAccountsRow } from './ChartOfAccountsRow'
