+++ src/components/ChartOfAccountsTable/index.tsx
@@ -1,1 +1,1 @@
-export { ChartOfAccountsTable } from './ChartOfAccountsTable'
+export { ChartOfAccountsTableWithPanel as ChartOfAccountsTable } from './ChartOfAccountsTableWithPanel'
