+++ src/styles/chart_of_accounts.scss
@@ -8,84 +8,16 @@
   padding: var(--spacing-2xl);
 }
 
-.Layer__chart-of-accounts__table {
-  width: 100%;
-  border-collapse: collapse;
-}
-
-.Layer__chart-of-accounts__actions {
-  display: flex;
-  align-items: center;
-  gap: var(--spacing-sm);
-}
-
-.Layer__table-cell.Layer__coa__name,
-.Layer__table-cell.Layer__coa__balance {
-  color: var(--color-base-800);
-}
-
-.Layer__table-header.Layer__coa__name {
-  padding-left: var(--spacing-xl);
-}
-
-.Layer__table-cell.Layer__coa__name {
-  width: calc(100% - 192px - 50%);
-}
-
-.Layer__table-cell.Layer__coa__type,
-.Layer__table-cell.Layer__coa__subtype {
-  width: calc(100% - 192px - 75%);
-}
-
-.Layer__table-cell.Layer__coa__balance {
-  font-variation-settings: 'wght' var(--font-weight-bold);
-}
-
-.Layer__chart-of-accounts__table
-  .Layer__table-row--depth-0
-  .Layer__table-cell.Layer__coa__name {
-  font-variation-settings: 'wght' var(--font-weight-bold);
-}
-
-.Layer__coa__balance {
-  text-align: right;
-  justify-content: flex-end;
-}
-
-.Layer__table-cell.Layer__coa__actions {
-  width: 160px;
-
-  .Layer__table-cell-content {
-    display: flex;
-    justify-content: flex-end;
-    align-items: center;
-    padding-top: var(--spacing-3xs);
-    padding-bottom: var(--spacing-3xs);
-    padding-right: var(--spacing-xl);
-    gap: var(--spacing-xs);
-
-    & > .Layer__btn {
-      opacity: 0;
-      visibility: hidden;
-      transition: all 300ms ease-out;
-      transform: scale(0.92);
-    }
-  }
-
-  .Layer__btn--icon-only {
-    width: 32px;
-    height: 32px;
-    min-height: 32px;
-    max-height: 32px;
-    padding: 0;
-    box-sizing: border-box;
-  }
-}
-
-.Layer__chart-of-accounts__table
-  .Layer__table-row:hover
-  .Layer__coa__actions
+.Layer__coa__actions {
   .Layer__btn {
+    opacity: 0;
+    visibility: hidden;
+    transition: all 300ms ease-out;
+    transform: scale(0.92);
+  }
+}
+
+.Layer__table-row:hover .Layer__coa__actions .Layer__btn {
   opacity: 1;
   visibility: visible;
   transform: scale(1);
@@ -117,40 +49,6 @@
   align-items: center;
   justify-content: flex-end;
   padding-top: var(--spacing-xl);
-}
-
-.Layer__chart-of-accounts__table .Layer__table-row {
-  &:not(.Layer__table-row--header) {
-    cursor: pointer;
-  }
-
-  &:not(.Layer__table-row--header):hover {
-    background-color: var(--color-base-50);
-  }
-}
-
-.Layer__table__expand-icon {
-  transition: transform 150ms ease-out;
-  color: var(--color-base-600);
-  margin-left: -15px;
-  padding: var(--spacing-sm);
-}
-
-.Layer__chart-of-accounts__table
-  .Layer__table-row:not(.Layer__table-row--depth-0):not(
-    .Layer__table-row--header
-  ) {
-  background-color: var(--color-base-50);
-
-  &:hover {
-    background-color: var(--color-base-100);
-  }
-}
-
-.Layer__chart-of-accounts__table
-  .Layer__table-row--collapsed
-  .Layer__table__expand-icon {
-  transform: rotate(-90deg);
 }
 
 .Layer__chart-of-accounts__form {
@@ -270,10 +168,6 @@
 }
 
 @container (max-width: 1024px) {
-  .Layer__table-cell.Layer__coa__actions .Layer__table-cell-content {
-    padding-right: var(--spacing-md);
-  }
-
   .Layer__chart-of-accounts__sidebar__header .actions {
     display: none;
   }
@@ -288,38 +182,10 @@
     .Layer__mobile--hidden {
       display: none;
     }
-
-    .Layer__desktop--hidden {
-      &.Layer__table-header {
-        display: table-cell;
-      }
-
-      &.Layer__table-cell-content {
-        display: flex;
-      }
-    }
-
-    .Layer__coa__type .Layer__table-cell-content {
-      display: flex;
-      flex-direction: column;
-      align-items: flex-start;
-
-      &.Layer__mobile--hidden {
-        display: none;
-      }
-    }
-
-    .Layer__table-cell.Layer__coa__actions {
-      width: 64px;
-    }
   }
 }
 
 @container (max-width: 760px) {
-  .Layer__table-cell.Layer__coa__actions {
-    width: 64px;
-  }
-
   .Layer__chart-of-accounts__row---mobile {
     visibility: visible;
     display: table-row;
